model/requirements.txt
# Image Processing
Pillow>=10.3.0
opencv-python-headless>=4.9.0.80
rembg>=2.0.56
numpy>=1.26.3

# Machine Learning
torch>=2.1.2
torchvision>=0.16.2
transformers>=4.37.0
open-clip-torch>=2.24.0
scikit-learn>=1.4.0
